% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_thresh.R
\name{get.int}
\alias{get.int}
\title{Get thresholds from U matrix}
\usage{
get.int(x, kstar, trt.code, trt.sub)
}
\arguments{
\item{x}{Column of \eqn{U} matrix, containing all possible threshold
solutions for a data point.}

\item{kstar}{Base-case optimal treatment.}

\item{trt.code}{Vector of (possibly recoded) treatments. See
\code{nma_thresh} parameter of the same name.}

\item{trt.sub}{Vector of treatment indices, subset of trt.code, to consider.
See \code{nma_thresh} parameter of the same name.}
}
\value{
Data frame of thresholds and new optimal treatments with columns
  \code{lo}, \code{lo.newkstar}, \code{hi}, and \code{hi.newkstar}.
}
\description{
This function is intended for internal use only, and is called by
\code{nma_thresh} automatically.
}

