% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SocAnx.post.cov}
\alias{SocAnx.post.cov}
\title{Posterior covariance matrix from Social Anxiety NMA}
\format{An object of class \code{matrix} with 186 rows and 186 columns.}
\source{
Generated from WinBUGS output, using the WinBUGS code from
  Mayo-Wilson et al. (2014). See also \code{vignette("Examples", package =
  "nmathresh")}.
}
\usage{
SocAnx.post.cov
}
\description{
The posterior covariance matrix of the variables \code{d} (basic treatment
effect parameters) and \code{delta} (shrunken random effects estimates for
each study).
}
\references{
Mayo-Wilson E, Dias S, Mavranezouli I, Kew K, Clark DM, Ades AE,
  et al. Psychological and pharmacological interventions for social anxiety
  disorder in adults: a systematic review and network meta-analysis. Lancet
  Psychiatry 2014;1:368-76.
  http://dx.doi.org/10.1016/S2215-0366(14)70329-3
}
\seealso{
\code{\link{SocAnx.post.summary}}
}
\keyword{datasets}
