% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Thrombo.post.cov}
\alias{Thrombo.post.cov}
\title{Posterior covariance matrix from Thrombolytics NMA}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 5 columns.
}
\source{
Generated from WinBUGS output, using the WinBUGS code from Caldwell
  et al. (2005). See also \code{vignette("Examples", package = "nmathresh")}.
}
\usage{
Thrombo.post.cov
}
\description{
The posterior covariance matrix of the basic treatment effect parameters.
}
\references{
Caldwell DM, Ades AE, Higgins JPT. Simultaneous comparison of
  multiple treatments: combining direct and indirect evidence. Brit Med J
  2005;331:897-900. http://dx.doi.org/10.1136/bmj.331.7521.897
}
\seealso{
\code{\link{Thrombo.post.summary}}
}
\keyword{datasets}
