% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Thrombo.post.summary}
\alias{Thrombo.post.summary}
\title{Posterior summary from Thrombolytics NMA}
\format{A \code{summary.mcmc} object. The key components for our use are:
  \describe{ \item{statistics}{Matrix containing the posterior summary
  statistics, with columns for \code{Mean}, \code{SD}, \code{Naive SE}, and
  \code{Time-series SE} (also known as the Monte-Carlo standard error)}
  \item{quantiles}{Matrix containing the posterior 2.5\%, 25\%, 50\%, 75\%,
  and 97.5\% quantiles}}}
\source{
Generated from WinBUGS output, using the WinBUGS code from Caldwell
  et al. (2005). See also \code{vignette("Examples", package = "nmathresh")}.
}
\usage{
Thrombo.post.summary
}
\description{
A \code{summary.mcmc} object of the type produced by the \code{coda} package,
containing the requisite posterior summary information on the variables
\code{dd}, the contrasts of the treatment effect parameters.
}
\references{
Caldwell DM, Ades AE, Higgins JPT. Simultaneous comparison of
  multiple treatments: combining direct and indirect evidence. Brit Med J
  2005;331:897-900. http://dx.doi.org/10.1136/bmj.331.7521.897
}
\seealso{
\code{\link[coda]{summary.mcmc}}, \code{\link{Thrombo.post.cov}}
}
\keyword{datasets}
